исправляет ошибки, добавляет модуль
