+++ index.html
@@ -115,8 +115,8 @@
 
                 <!-- Добавление хэш-тегов и комментария к изображению -->
                 <fieldset class="img-upload__text text">
-                  <input class="text__hashtags" type="text" name="hashtags" id="input-hashtag" placeholder="#ХэшТег" minlength="2" maxlength="104">
-                  <textarea class="text__description" name="description" id="input-comment" form="upload-select-image" placeholder="Ваш комментарий..." minlength="1" maxlength="140"></textarea>
+                  <input class="text__hashtags" type="text" name="hashtags" id="input-hashtag" placeholder="#ХэшТег">
+                  <textarea class="text__description" name="description" id="input-comment" form="upload-select-image" placeholder="Ваш комментарий..."></textarea>
                 </fieldset>
 
                 <!-- Кнопка для отправки данных на сервер -->
