+++ js/main.js
@@ -1,2 +1,3 @@
 import './picture-painting.js';
 import './popup-editor.js';
+import './input-validation.js';
